feat: Added localization strings
