+++ Core/Utils/AppConvigVariables.struct.swift
@@ -8,4 +8,5 @@ struct AppConfigVariables {
     static let arrivalsCode: String = ProcessInfo.processInfo.environment["ARRIVALS_CODE"] ?? ""
     static let topTrendsCode: String = ProcessInfo.processInfo.environment["TOP_TRENDS_CODE"] ?? ""
     static let recommendationCode: String = ProcessInfo.processInfo.environment["RECOMMENDATION_CODE"] ?? ""
+    static let inAppNotificationImageUrl: String = ProcessInfo.processInfo.environment["IN_APP_NOTIFICATION_IMAGE_URL"] ?? ""
 }
